trip/serializers.py
from rest_framework import serializers

class TripInputSerializer(serializers.Serializer):
    current_location = serializers.CharField()
    pickup_location = serializers.CharField()
    dropoff_location = serializers.CharField()
    cycle_used = serializers.IntegerField()
